collecting gems and file structure
