+++ artemis/constants.py
@@ -1,0 +1,12 @@
+ASSETS = 'assets/'
+
+BLOCKS_Y = 20
+BLOCKS_X = 40
+SCALING = 0.25
+HEIGHT = int(BLOCKS_Y * SCALING * 128)
+WIDTH = int(BLOCKS_X * SCALING * 128)
+SIDE = int(128 * SCALING)
+TOP = int(600 * SCALING)
+
+SPEED = 20 * SCALING
+GRAVITY = 10 * SCALING
