+++ artemis/main.py
@@ -1,0 +1,4 @@
+from game import Game
+
+
+Game()
